requirements.txt
streamlit==1.32.2
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
